refactor: restructured auth
