+++ client/internal/auth/token/token.go
@@ -1,4 +1,4 @@
-package auth
+package token
 
 import (
 	"crypto/rand"
